+++ test/h5file/h5portable.cpp
@@ -11,7 +11,7 @@
 #include "TestRunner.hpp"
 #include "ReadHDF5.hpp"
 #include "MBTagConventions.hpp"
-#include "FileOptions.hpp"
+#include "moab/FileOptions.hpp"
 #include <vector>
 #include <stdlib.h>
 #include <iostream>
@@ -107,7 +107,7 @@ void create_mesh( const char* filename )
   
   
   for (size_t i = 0; i < NUM_VERT; ++i) {
-    double coords[] = { i % 9, i / 9, Z };
+    double coords[] = { static_cast<double>(i % 9), static_cast<double>(i / 9), static_cast<double>(Z) };
     rval = mb.create_vertex( coords, verts[i] );
     CHECK_ERR(rval);
   }
